refactor: updated Footer component path, place common components in new folder
