+++ src/components/common/Footer.tsx
@@ -1,7 +1,7 @@
 import { FaFacebook, FaGoogle, FaTwitter, FaYoutube } from 'react-icons/fa';
 import { Link } from 'react-router-dom';
-import Logo from '../assets/Logo/Logo-Full-Light.png';
-import { FooterLink2 } from '../data/footer-links';
+import Logo from '../../assets/Logo/Logo-Full-Light.png';
+import { FooterLink2 } from '../../data/footer-links';
 
 const BottomFooter = ['Privacy Policy', 'Cookie Policy', 'Terms'];
 const Resources = [
